pyproject.toml
[tool.poetry]
name = "power-consumption"
version = "0.1.0"
description = ""
authors = ["false <false@wizard-limit.net>"]

[tool.poetry.dependencies]
python = "^3.7.1"
pyserial = {version = "^3.5", optional = true}
psycopg2-binary = "^2.8.6"
bokeh = {version = "^2.2.3", optional = true}
mh-z19 = {version = "^3.0.2", optional = true}
pandas = {version = "^1.2.4", optional = true}
smbus = {version = "^1.1.post2", optional = true}

[tool.poetry.extras]
poller = ["pyserial", "mh-z19", "smbus"]
graph = ["bokeh", "pandas"]

[tool.poetry.dev-dependencies]
black = "^20.8b1"
flake8 = "^3.8.4"
flake8-docstrings = "^1.5.0"
mypy = "^0.800"
ipython = "^7.20.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120
